requirements.txt
flask
requests
pinecone-client
langchain
pymupdf
python-dotenv
langchain-community
langchain_pinecone
